+++ app/src/main/java/ru/sandbox/androidacademyapp/MoviesAdapter.kt
@@ -10,9 +10,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
-import ru.sandbox.androidacademyapp.api.MovieResponse
 import ru.sandbox.androidacademyapp.data.db.entites.Movie
-import kotlin.math.roundToInt
 
 class MoviesAdapter(private val clickListener: OnRecyclerItemClicked): RecyclerView.Adapter<MovieViewHolder>() {
