+++ app/src/main/java/ru/sandbox/androidacademyapp/MoviesViewModel.kt
@@ -4,7 +4,7 @@ import ru.sandbox.androidacademyapp.util.SingleLiveEvent
 import androidx.lifecycle.*
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
-import ru.sandbox.androidacademyapp.api.Result
+import ru.sandbox.androidacademyapp.data.api.Result
 import ru.sandbox.androidacademyapp.data.db.entites.Movie
 import ru.sandbox.androidacademyapp.repository.IMovieRepository
