+++ app/src/main/java/ru/sandbox/androidacademyapp/data/api/MoviesApiHeaderInterceptor.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.api
+package ru.sandbox.androidacademyapp.data.api
 
 import okhttp3.Interceptor
 import okhttp3.Response
