+++ app/src/main/java/ru/sandbox/androidacademyapp/repository/MovieRepository.kt
@@ -3,10 +3,10 @@ package ru.sandbox.androidacademyapp.repository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import ru.sandbox.androidacademyapp.BuildConfig
-import ru.sandbox.androidacademyapp.api.MoviesApi
-import ru.sandbox.androidacademyapp.api.ActorResponse
-import ru.sandbox.androidacademyapp.api.MovieResponse
-import ru.sandbox.androidacademyapp.api.Result
+import ru.sandbox.androidacademyapp.data.api.MoviesApi
+import ru.sandbox.androidacademyapp.data.api.ActorResponse
+import ru.sandbox.androidacademyapp.data.api.MovieResponse
+import ru.sandbox.androidacademyapp.data.api.Result
 import ru.sandbox.androidacademyapp.data.db.MoviesDao
 import ru.sandbox.androidacademyapp.data.db.entites.Actor
 import ru.sandbox.androidacademyapp.data.db.entites.Movie
